# project/pwa-backend-master/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "HEARBK Backend",
    "description": "HearBK ",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Register a new user",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/RegisterUser"
            }
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/SuccessMessage"
            }
          }
        }
      }
    },
...
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "Update details of a new user",
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "schema": {
              "$ref": "#/definitions/AuthTokenHeader"
            }
          },
          {
            "name": "user",
            "in": "body",
            "description": "User login api",
            "schema": {
              "$ref": "#/definitions/RegisterUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token details",
            "schema": {
              "$ref": "#/definitions/SuccessMessage"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "RegisterUser": {
      "type": "object",
      "required": [
        "email",
        "user_name",
        "display_name",
        "password"
      ],
      "properties": {
        "user_name": {
          "type": "string",
          "uniqueItems": true
        },
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "display_name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "SuccessMessage": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "AuthTokenHeader": {
      "type": "string",
      "example": "<access_token>"
    },
...
  }
}